Fix LB issues

+++ resource_obmcs_loadbalancer_loadbalancer.go
@@ -6,6 +6,8 @@ import (
 	"github.com/MustWin/baremetal-sdk-go"
 	"github.com/hashicorp/terraform/helper/schema"
 
+	"strings"
+
 	"github.com/oracle/terraform-provider-baremetal/client"
 	"github.com/oracle/terraform-provider-baremetal/crud"
 )
@@ -228,6 +230,10 @@ func (s *LoadBalancerResourceCrud) SetData() {
 
 // Delete makes a request to delete the load balancer
 func (s *LoadBalancerResourceCrud) Delete() (e error) {
+	// TODO: make sure this actually works
+	if strings.Contains(s.D.Id(), "ocid1.loadbalancerworkrequest") {
+		return
+	}
 	var workReqID string
 	workReqID, e = s.Client.DeleteLoadBalancer(s.D.Id(), nil)
 	if e != nil {
